requirements.txt
requests>=2.28.1
pytest>=7.2.0
pytest-bdd>=6.1.1
allure-pytest-bdd>=2.9.43
playwright>=1.8.0
pytest-xdist>=3.1.0